+++ tests/ecmascript/test-bi-array-proto-concatp-trailing-gaps.js
@@ -56,13 +56,10 @@
  *
  *  The default behavior for Duktape will be changed to match the other engines,
  *  even though this seems non-standard.
+ *
+ *  ES2015 fixed this specification "bug" so the non-standard behavior is
+ *  now standard.
  */
-
-/*---
-{
-    "nonstandard": true
-}
----*/
 
 /*===
 3 1 2 3